requirements.txt
numpy
torch
matplotlib
tqdm
pygad
pyyaml
scipy
scikit-learn
pandas
pillow
h5py

